Updated directory structure. No file in a child directory depends on a file in the parent directory. Made the mark_message_as_read function asynchronous.
